+++ tests/JwtAuthenticationTest.php
@@ -1,17 +1,33 @@
 <?php
 
+/*
+
+Copyright (c) 2015-2019 Mika Tuupola
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
+
+*/
+
 /**
- * This file is part of PSR-7 & PSR-15 JWT Authentication middleware
- *
- * Copyright (c) 2015-2018 Mika Tuupola
- *
- * Licensed under the MIT license:
- *   http://www.opensource.org/licenses/mit-license.php
- *
- * Project home:
- *   https://github.com/tuupola/slim-jwt-auth
- *   https://appelsiini.net/projects/slim-jwt-auth
- *
+ * @see       https://github.com/tuupola/slim-jwt-auth
+ * @see       https://appelsiini.net/projects/slim-jwt-auth
+ * @license   https://www.opensource.org/licenses/mit-license.php
  */
 
 namespace Tuupola\Middleware;
@@ -29,17 +45,27 @@ use Tuupola\Middleware\JwtAuthentication\RequestPathRule;
 class JwtAuthenticationTest extends TestCase
 {
     /* @codingStandardsIgnoreStart */
-    public static $token = "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJBY21lIFRvb3RocGljcyBMdGQiLCJpYXQiOjE0Mjg4MTk5NDEsImV4cCI6MTc0NDM1Mjc0MSwiYXVkIjoid3d3LmV4YW1wbGUuY29tIiwic3ViIjoic29tZW9uZUBleGFtcGxlLmNvbSIsInNjb3BlIjpbInJlYWQiLCJ3cml0ZSIsImRlbGV0ZSJdfQ.YzPxtyHLqiJMUaPE6DzBonGUyqLlddxIisxSFk2Gk7Y";
+    public static $acmeToken = "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiIsImtpZCI6ImFjbWUifQ.eyJpc3MiOiJBY21lIFRvb3RocGljcyBMdGQiLCJpYXQiOiIxNDI4ODE5OTQxIiwiZXhwIjoiMTc0NDM1Mjc0MSIsImF1ZCI6Ind3dy5leGFtcGxlLmNvbSIsInN1YiI6InNvbWVvbmVAZXhhbXBsZS5jb20iLCJzY29wZSI6WyJyZWFkIiwid3JpdGUiLCJkZWxldGUiXX0.yBhYlsMabKTh31taAiH8i2ScPMKm84jxIDNxft6EiTA";
+    public static $betaToken = "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiIsImtpZCI6ImJldGEifQ.eyJraWQiOiJiZXRhIiwiaXNzIjoiQmV0YSBTcG9uc29yc2hpcCBMdGQiLCJpYXQiOiIxNDI4ODE5OTQxIiwiZXhwIjoiMTc0NDM1Mjc0MSIsImF1ZCI6Ind3dy5leGFtcGxlLmNvbSIsInN1YiI6InNvbWVvbmVAZXhhbXBsZS5jb20iLCJzY29wZSI6WyJyZWFkIl19.msxcBx4_ZQtCkkjHyTDWDC0mac4cFNSxLqkzNL30JB8";
     public static $expired = "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJBY21lIFRvb3RocGljcyBMdGQiLCJpYXQiOjE0Mjg4MTk5NDEsImV4cCI6MTQ4MDcyMzIwMCwiYXVkIjoid3d3LmV4YW1wbGUuY29tIiwic3ViIjoic29tZW9uZUBleGFtcGxlLmNvbSIsInNjb3BlIjpbInJlYWQiLCJ3cml0ZSIsImRlbGV0ZSJdfQ.ZydGEHVmca4ofQRCuMOfZrUXprAoe5GcySg4I-lwIjc";
     /* @codingStandardsIgnoreEnd */
 
-    public static $token_as_array = [
+    public static $acmeTokenArray = [
         "iss" => "Acme Toothpics Ltd",
         "iat" => "1428819941",
         "exp" => "1744352741",
         "aud" => "www.example.com",
         "sub" => "someone@example.com",
         "scope" => ["read", "write", "delete"]
+    ];
+
+    public static $betaTokenArray = [
+        "iss" => "Beta Sponsorship Ltd",
+        "iat" => "1428819941",
+        "exp" => "1744352741",
+        "aud" => "www.example.com",
+        "sub" => "someone@example.com",
+        "scope" => ["read"]
     ];
 
     public function testShouldBeTrue()
@@ -74,7 +100,7 @@ class JwtAuthenticationTest extends TestCase
     {
         $request = (new ServerRequestFactory)
             ->createServerRequest("GET", "https://example.com/api")
-            ->withHeader("X-Token", "Bearer " . self::$token);
+            ->withHeader("X-Token", "Bearer " . self::$acmeToken);
 
         $default = function (ServerRequestInterface $request) {
             $response = (new ResponseFactory)->createResponse();
@@ -99,7 +125,7 @@ class JwtAuthenticationTest extends TestCase
     {
         $request = (new ServerRequestFactory)
             ->createServerRequest("GET", "https://example.com/api")
-            ->withHeader("X-Token", self::$token);
+            ->withHeader("X-Token", self::$acmeToken);
 
         $default = function (ServerRequestInterface $request) {
             $response = (new ResponseFactory)->createResponse();
@@ -125,7 +151,7 @@ class JwtAuthenticationTest extends TestCase
     {
         $request = (new ServerRequestFactory)
             ->createServerRequest("GET", "https://example.com/api")
-            ->withCookieParams(["nekot" => self::$token]);
+            ->withCookieParams(["nekot" => self::$acmeToken]);
 
         $default = function (ServerRequestInterface $request) {
             $response = (new ResponseFactory)->createResponse();
@@ -146,11 +172,143 @@ class JwtAuthenticationTest extends TestCase
         $this->assertEquals("Success", $response->getBody());
     }
 
-    public function testShouldAlterResponseWithAfter()
+    public function testShouldReturn200WithTokenFromBearerCookie()
     {
         $request = (new ServerRequestFactory)
             ->createServerRequest("GET", "https://example.com/api")
-            ->withHeader("Authorization", "Bearer " . self::$token);
+            ->withCookieParams(["nekot" => "Bearer " . self::$acmeToken]);
+
+        $default = function (ServerRequestInterface $request) {
+            $response = (new ResponseFactory)->createResponse();
+            $response->getBody()->write("Success");
+            return $response;
+        };
+
+        $collection = new MiddlewareCollection([
+            new JwtAuthentication([
+                "secret" => "supersecretkeyyoushouldnotcommittogithub",
+                "cookie" => "nekot",
+            ])
+        ]);
+
+        $response = $collection->dispatch($request, $default);
+
+        $this->assertEquals(200, $response->getStatusCode());
+        $this->assertEquals("Success", $response->getBody());
+    }
+
+
+    public function testShouldReturn200WithSecretArray()
+    {
+        $request = (new ServerRequestFactory)
+            ->createServerRequest("GET", "https://example.com/api")
+            ->withHeader("Authorization", "Bearer " . self::$betaToken);
+
+        $default = function (ServerRequestInterface $request) {
+            $response = (new ResponseFactory)->createResponse();
+            $response->getBody()->write("Success");
+            return $response;
+        };
+
+        $collection = new MiddlewareCollection([
+            new JwtAuthentication([
+                "secret" => [
+                    "acme" =>"supersecretkeyyoushouldnotcommittogithub",
+                    "beta" =>"anothersecretkeyfornevertocommittogithub"
+                ]
+            ])
+        ]);
+
+        $response = $collection->dispatch($request, $default);
+        $this->assertEquals(200, $response->getStatusCode());
+        $this->assertEquals("Success", $response->getBody());
+    }
+
+    public function testShouldReturn401WithSecretArray()
+    {
+        $request = (new ServerRequestFactory)
+            ->createServerRequest("GET", "https://example.com/api")
+            ->withHeader("Authorization", "Bearer " . self::$betaToken);
+
+        $default = function (ServerRequestInterface $request) {
+            $response = (new ResponseFactory)->createResponse();
+            $response->getBody()->write("Success");
+            return $response;
+        };
+
+        $collection = new MiddlewareCollection([
+            new JwtAuthentication([
+                "secret" => [
+                    "xxxx" =>"supersecretkeyyoushouldnotcommittogithub",
+                    "yyyy" =>"anothersecretkeyfornevertocommittogithub"
+                ]
+            ])
+        ]);
+
+        $response = $collection->dispatch($request, $default);
+        $this->assertEquals(401, $response->getStatusCode());
+        $this->assertEquals("", $response->getBody());
+    }
+
+    public function testShouldReturn200WithSecretArrayAccess()
+    {
+        $request = (new ServerRequestFactory)
+            ->createServerRequest("GET", "https://example.com/api")
+            ->withHeader("Authorization", "Bearer " . self::$betaToken);
+
+        $default = function (ServerRequestInterface $request) {
+            $response = (new ResponseFactory)->createResponse();
+            $response->getBody()->write("Success");
+            return $response;
+        };
+
+        $secret = new ArrayAccessImpl();
+        $secret["acme"] = "supersecretkeyyoushouldnotcommittogithub";
+        $secret["beta"] ="anothersecretkeyfornevertocommittogithub";
+
+        $collection = new MiddlewareCollection([
+            new JwtAuthentication([
+                "secret" => $secret
+            ])
+        ]);
+
+        $response = $collection->dispatch($request, $default);
+        $this->assertEquals(200, $response->getStatusCode());
+        $this->assertEquals("Success", $response->getBody());
+    }
+
+    public function testShouldReturn401WithSecretArrayAccess()
+    {
+        $request = (new ServerRequestFactory)
+            ->createServerRequest("GET", "https://example.com/api")
+            ->withHeader("Authorization", "Bearer " . self::$betaToken);
+
+        $default = function (ServerRequestInterface $request) {
+            $response = (new ResponseFactory)->createResponse();
+            $response->getBody()->write("Success");
+            return $response;
+        };
+
+        $secret = new ArrayAccessImpl();
+        $secret["xxxx"] = "supersecretkeyyoushouldnotcommittogithub";
+        $secret["yyyy"] = "anothersecretkeyfornevertocommittogithub";
+
+        $collection = new MiddlewareCollection([
+            new JwtAuthentication([
+                "secret" => $secret
+            ])
+        ]);
+
+        $response = $collection->dispatch($request, $default);
+        $this->assertEquals(401, $response->getStatusCode());
+        $this->assertEquals("", $response->getBody());
+    }
+
+    public function testShouldAlterResponseWithAnonymousAfter()
+    {
+        $request = (new ServerRequestFactory)
+            ->createServerRequest("GET", "https://example.com/api")
+            ->withHeader("Authorization", "Bearer " . self::$acmeToken);
 
         $default = function (ServerRequestInterface $request) {
             $response = (new ResponseFactory)->createResponse();
@@ -173,11 +331,67 @@ class JwtAuthenticationTest extends TestCase
         $this->assertEquals("plants crave", (string) $response->getHeaderLine("X-Brawndo"));
     }
 
+    public function testShouldAlterResponseWithInvokableAfter()
+    {
+        $request = (new ServerRequestFactory)
+            ->createServerRequest("GET", "https://example.com/api")
+            ->withHeader("Authorization", "Bearer " . self::$acmeToken);
+
+        $default = function (ServerRequestInterface $request) {
+            $response = (new ResponseFactory)->createResponse();
+            $response->getBody()->write("Success");
+            return $response;
+        };
+
+        $collection = new MiddlewareCollection([
+            new JwtAuthentication([
+                "secret" => "supersecretkeyyoushouldnotcommittogithub",
+                "after" => new TestAfterHandler
+            ])
+        ]);
+
+        $response = $collection->dispatch($request, $default);
+
+        $this->assertEquals(200, $response->getStatusCode());
+        $this->assertEquals(
+            "plants crave",
+            (string) $response->getHeaderLine("X-Brawndo")
+        );
+    }
+
+    public function testShouldAlterResponseWithArrayNotationAfter()
+    {
+        $request = (new ServerRequestFactory)
+            ->createServerRequest("GET", "https://example.com/api")
+            ->withHeader("Authorization", "Bearer " . self::$acmeToken);
+
+        $default = function (ServerRequestInterface $request) {
+            $response = (new ResponseFactory)->createResponse();
+            $response->getBody()->write("Success");
+            return $response;
+        };
+
+        $collection = new MiddlewareCollection([
+            new JwtAuthentication([
+                "secret" => "supersecretkeyyoushouldnotcommittogithub",
+                "after" => [TestAfterHandler::class, "after"]
+            ])
+        ]);
+
+        $response = $collection->dispatch($request, $default);
+
+        $this->assertEquals(200, $response->getStatusCode());
+        $this->assertEquals(
+            "like from toilet?",
+            (string) $response->getHeaderLine("X-Water")
+        );
+    }
+
     public function testShouldReturn401WithInvalidAlgorithm()
     {
         $request = (new ServerRequestFactory)
             ->createServerRequest("GET", "https://example.com/api")
-            ->withHeader("Authorization", "Bearer " . self::$token);
+            ->withHeader("Authorization", "Bearer " . self::$acmeToken);
 
         $default = function (ServerRequestInterface $request) {
             $response = (new ResponseFactory)->createResponse();
@@ -226,7 +440,7 @@ class JwtAuthenticationTest extends TestCase
     {
         $request = (new ServerRequestFactory)
             ->createServerRequest("GET", "https://example.com/api")
-            ->withHeader("Authorization", "Bearer invalid" . self::$token);
+            ->withHeader("Authorization", "Bearer invalid" . self::$acmeToken);
 
         $default = function (ServerRequestInterface $request) {
             $response = (new ResponseFactory)->createResponse();
@@ -325,7 +539,7 @@ class JwtAuthenticationTest extends TestCase
 
         $request = (new ServerRequestFactory)
             ->createServerRequest("GET", "http://example.com/api")
-            ->withHeader("Authorization", "Bearer " . self::$token);
+            ->withHeader("Authorization", "Bearer " . self::$acmeToken);
 
         $default = function (ServerRequestInterface $request) {
             $response = (new ResponseFactory)->createResponse();
@@ -346,7 +560,7 @@ class JwtAuthenticationTest extends TestCase
     {
         $request = (new ServerRequestFactory)
             ->createServerRequest("GET", "http://example.com/api")
-            ->withHeader("Authorization", "Bearer " . self::$token);
+            ->withHeader("Authorization", "Bearer " . self::$acmeToken);
 
         $default = function (ServerRequestInterface $request) {
             $response = (new ResponseFactory)->createResponse();
@@ -371,7 +585,7 @@ class JwtAuthenticationTest extends TestCase
     {
         $request = (new ServerRequestFactory)
             ->createServerRequest("GET", "http://localhost/api")
-            ->withHeader("Authorization", "Bearer " . self::$token);
+            ->withHeader("Authorization", "Bearer " . self::$acmeToken);
 
         $default = function (ServerRequestInterface $request) {
             $response = (new ResponseFactory)->createResponse();
@@ -395,7 +609,7 @@ class JwtAuthenticationTest extends TestCase
     {
         $request = (new ServerRequestFactory)
             ->createServerRequest("GET", "http://example.com/api")
-            ->withHeader("Authorization", "Bearer " . self::$token);
+            ->withHeader("Authorization", "Bearer " . self::$acmeToken);
 
         $default = function (ServerRequestInterface $request) {
             $response = (new ResponseFactory)->createResponse();
@@ -420,13 +634,13 @@ class JwtAuthenticationTest extends TestCase
     {
         $request = (new ServerRequestFactory)
             ->createServerRequest("GET", "https://example.com/api")
-            ->withHeader("Authorization", "Bearer " . self::$token);
+            ->withHeader("Authorization", "Bearer " . self::$acmeToken);
 
         $default = function (ServerRequestInterface $request) {
-            $token = $request->getAttribute("token");
+            $acmeToken = $request->getAttribute("token");
 
             $response = (new ResponseFactory)->createResponse();
-            $response->getBody()->write($token["iss"]);
+            $response->getBody()->write($acmeToken["iss"]);
 
             return $response;
         };
@@ -447,13 +661,13 @@ class JwtAuthenticationTest extends TestCase
     {
         $request = (new ServerRequestFactory)
             ->createServerRequest("GET", "https://example.com/api")
-            ->withHeader("Authorization", "Bearer " . self::$token);
+            ->withHeader("Authorization", "Bearer " . self::$acmeToken);
 
         $default = function (ServerRequestInterface $request) {
-            $token = $request->getAttribute("nekot");
+            $acmeToken = $request->getAttribute("nekot");
 
             $response = (new ResponseFactory)->createResponse();
-            $response->getBody()->write($token["iss"]);
+            $response->getBody()->write($acmeToken["iss"]);
 
             return $response;
         };
@@ -471,13 +685,14 @@ class JwtAuthenticationTest extends TestCase
         $this->assertEquals("Acme Toothpics Ltd", $response->getBody());
     }
 
-    public function testShouldCallAfter()
+    public function testShouldCallAfterWithProperArguments()
     {
         $request = (new ServerRequestFactory)
             ->createServerRequest("GET", "https://example.com/api")
-            ->withHeader("Authorization", "Bearer " . self::$token);
+            ->withHeader("Authorization", "Bearer " . self::$acmeToken);
 
-        $dummy = null;
+        $decoded = null;
+        $token = null;
 
         $default = function (ServerRequestInterface $request) {
             $response = (new ResponseFactory)->createResponse();
@@ -488,8 +703,9 @@ class JwtAuthenticationTest extends TestCase
         $collection = new MiddlewareCollection([
             new JwtAuthentication([
                 "secret" => "supersecretkeyyoushouldnotcommittogithub",
-                "after" => function ($response, $arguments) use (&$dummy) {
-                    $dummy = $arguments["decoded"];
+                "after" => function ($response, $arguments) use (&$decoded, &$token) {
+                    $decoded = $arguments["decoded"];
+                    $token = $arguments["token"];
                 }
             ])
         ]);
@@ -498,10 +714,44 @@ class JwtAuthenticationTest extends TestCase
 
         $this->assertEquals(200, $response->getStatusCode());
         $this->assertEquals("Success", $response->getBody());
-        $this->assertEquals(self::$token_as_array, (array) $dummy);
+        $this->assertEquals(self::$acmeTokenArray, (array) $decoded);
+        $this->assertEquals(self::$acmeToken, $token);
     }
 
-    public function testShouldCallError()
+    public function testShouldCallBeforeWithProperArguments()
+    {
+        $request = (new ServerRequestFactory)
+            ->createServerRequest("GET", "https://example.com/api")
+            ->withHeader("Authorization", "Bearer " . self::$acmeToken);
+
+        $decoded = null;
+        $token = null;
+
+        $default = function (ServerRequestInterface $request) {
+            $response = (new ResponseFactory)->createResponse();
+            $response->getBody()->write("Success");
+            return $response;
+        };
+
+        $collection = new MiddlewareCollection([
+            new JwtAuthentication([
+                "secret" => "supersecretkeyyoushouldnotcommittogithub",
+                "before" => function ($response, $arguments) use (&$decoded, &$token) {
+                    $decoded = $arguments["decoded"];
+                    $token = $arguments["token"];
+                }
+            ])
+        ]);
+
+        $response = $collection->dispatch($request, $default);
+
+        $this->assertEquals(200, $response->getStatusCode());
+        $this->assertEquals("Success", $response->getBody());
+        $this->assertEquals(self::$acmeTokenArray, (array) $decoded);
+        $this->assertEquals(self::$acmeToken, $token);
+    }
+
+    public function testShouldCallAnonymousErrorFunction()
     {
         $request = (new ServerRequestFactory)
             ->createServerRequest("GET", "https://example.com/api");
@@ -528,6 +778,58 @@ class JwtAuthenticationTest extends TestCase
         $this->assertEquals(401, $response->getStatusCode());
         $this->assertEquals("", $response->getBody());
         $this->assertTrue($dummy);
+    }
+
+    public function testShouldCallInvokableErrorClass()
+    {
+        $request = (new ServerRequestFactory)
+            ->createServerRequest("GET", "https://example.com/api");
+
+        $dummy = null;
+
+        $default = function (ServerRequestInterface $request) {
+            $response = (new ResponseFactory)->createResponse();
+            $response->getBody()->write("Success");
+            return $response;
+        };
+
+        $collection = new MiddlewareCollection([
+            new JwtAuthentication([
+                "secret" => "supersecretkeyyoushouldnotcommit",
+                "error" => new TestErrorHandler
+            ])
+        ]);
+
+        $response = $collection->dispatch($request, $default);
+
+        $this->assertEquals(402, $response->getStatusCode());
+        $this->assertEquals(TestErrorHandler::class, $response->getBody());
+    }
+
+    public function testShouldCallArrayNotationError()
+    {
+        $request = (new ServerRequestFactory)
+            ->createServerRequest("GET", "https://example.com/api");
+
+        $dummy = null;
+
+        $default = function (ServerRequestInterface $request) {
+            $response = (new ResponseFactory)->createResponse();
+            $response->getBody()->write("Success");
+            return $response;
+        };
+
+        $collection = new MiddlewareCollection([
+            new JwtAuthentication([
+                "secret" => "supersecretkeyyoushouldnotcommit",
+                "error" => [TestErrorHandler::class, "error"]
+            ])
+        ]);
+
+        $response = $collection->dispatch($request, $default);
+
+        $this->assertEquals(418, $response->getStatusCode());
+        $this->assertEquals(TestErrorHandler::class, $response->getBody());
     }
 
     public function testShouldCallErrorAndModifyBody()
@@ -590,7 +892,7 @@ class JwtAuthenticationTest extends TestCase
     {
         $request = (new ServerRequestFactory)
             ->createServerRequest("GET", "https://example.com/api")
-            ->withHeader("Authorization", "Bearer " . self::$token);
+            ->withHeader("Authorization", "Bearer " . self::$acmeToken);
 
         $default = function (ServerRequestInterface $request) {
             $response = (new ResponseFactory)->createResponse();
@@ -615,11 +917,11 @@ class JwtAuthenticationTest extends TestCase
         $this->assertEquals("", $response->getBody());
     }
 
-    public function testShouldModifyRequestUsingBefore()
+    public function testShouldModifyRequestUsingAnonymousBefore()
     {
         $request = (new ServerRequestFactory)
             ->createServerRequest("GET", "https://example.com/")
-            ->withHeader("Authorization", "Bearer " . self::$token);
+            ->withHeader("Authorization", "Bearer " . self::$acmeToken);
 
         $default = function (ServerRequestInterface $request) {
             $response = (new ResponseFactory)->createResponse();
@@ -641,6 +943,58 @@ class JwtAuthenticationTest extends TestCase
 
         $this->assertEquals(200, $response->getStatusCode());
         $this->assertEquals("test", (string) $response->getBody());
+    }
+
+    public function testShouldModifyRequestUsingInvokableBefore()
+    {
+        $request = (new ServerRequestFactory)
+            ->createServerRequest("GET", "https://example.com/")
+            ->withHeader("Authorization", "Bearer " . self::$acmeToken);
+
+        $default = function (ServerRequestInterface $request) {
+            $response = (new ResponseFactory)->createResponse();
+            $test = $request->getAttribute("test");
+            $response->getBody()->write($test);
+            return $response;
+        };
+
+        $collection = new MiddlewareCollection([
+            new JwtAuthentication([
+                "secret" => "supersecretkeyyoushouldnotcommittogithub",
+                "before" => new TestBeforeHandler
+            ])
+        ]);
+
+        $response = $collection->dispatch($request, $default);
+
+        $this->assertEquals(200, $response->getStatusCode());
+        $this->assertEquals("invoke", (string) $response->getBody());
+    }
+
+    public function testShouldModifyRequestUsingArrayNotationBefore()
+    {
+        $request = (new ServerRequestFactory)
+            ->createServerRequest("GET", "https://example.com/")
+            ->withHeader("Authorization", "Bearer " . self::$acmeToken);
+
+        $default = function (ServerRequestInterface $request) {
+            $response = (new ResponseFactory)->createResponse();
+            $test = $request->getAttribute("test");
+            $response->getBody()->write($test);
+            return $response;
+        };
+
+        $collection = new MiddlewareCollection([
+            new JwtAuthentication([
+                "secret" => "supersecretkeyyoushouldnotcommittogithub",
+                "before" => [TestBeforeHandler::class, "before"]
+            ])
+        ]);
+
+        $response = $collection->dispatch($request, $default);
+
+        $this->assertEquals(200, $response->getStatusCode());
+        $this->assertEquals("function", (string) $response->getBody());
     }
 
     public function testShouldHandleRulesArrayBug84()
@@ -683,11 +1037,43 @@ class JwtAuthenticationTest extends TestCase
         $this->assertEquals("Success", $response->getBody());
     }
 
+    public function testShouldHandleDefaultPathBug118()
+    {
+        $request = (new ServerRequestFactory)
+            ->createServerRequest("GET", "https://example.com/api");
+
+        $default = function (ServerRequestInterface $request) {
+            $response = (new ResponseFactory)->createResponse();
+            $response->getBody()->write("Success");
+            return $response;
+        };
+
+        $collection = new MiddlewareCollection([
+            new JwtAuthentication([
+                "secret" => "supersecretkeyyoushouldnotcommittogithub",
+                "ignore" => "/api/login",
+            ])
+        ]);
+
+        $response = $collection->dispatch($request, $default);
+
+        $this->assertEquals(401, $response->getStatusCode());
+        $this->assertEquals("", $response->getBody());
+
+        $request = (new ServerRequestFactory)
+            ->createServerRequest("GET", "https://example.com/api/login");
+
+        $response = $collection->dispatch($request, $default);
+
+        $this->assertEquals(200, $response->getStatusCode());
+        $this->assertEquals("Success", $response->getBody());
+    }
+
     public function testShouldBindToMiddleware()
     {
         $request = (new ServerRequestFactory)
             ->createServerRequest("GET", "https://example.com/")
-            ->withHeader("Authorization", "Bearer " . self::$token);
+            ->withHeader("Authorization", "Bearer " . self::$acmeToken);
 
         $default = function (ServerRequestInterface $request) {
             $response = (new ResponseFactory)->createResponse();
@@ -722,7 +1108,7 @@ class JwtAuthenticationTest extends TestCase
     {
         $request = (new ServerRequestFactory)
             ->createServerRequest("GET", "https://example.com/api")
-            ->withHeader("X-Token", "Bearer " . self::$token);
+            ->withHeader("X-Token", "Bearer " . self::$acmeToken);
 
         $response = (new ResponseFactory)->createResponse();
 
@@ -737,6 +1123,61 @@ class JwtAuthenticationTest extends TestCase
         };
 
         $response = $auth($request, $response, $next);
+
+        $this->assertEquals(200, $response->getStatusCode());
+        $this->assertEquals("Success", $response->getBody());
+    }
+
+    public function testShouldHaveUriInErrorHandlerIssue96()
+    {
+        $request = (new ServerRequestFactory)
+            ->createServerRequest("GET", "https://example.com/api/foo?bar=pop");
+
+        $dummy = null;
+
+        $default = function (ServerRequestInterface $request) {
+            $response = (new ResponseFactory)->createResponse();
+            $response->getBody()->write("Success");
+            return $response;
+        };
+
+        $collection = new MiddlewareCollection([
+            new JwtAuthentication([
+                "secret" => "supersecretkeyyoushouldnotcommittogithub",
+                "error" => function (ResponseInterface $response, $arguments) use (&$dummy) {
+                    $dummy = $arguments["uri"];
+                }
+            ])
+        ]);
+
+        $response = $collection->dispatch($request, $default);
+
+        $this->assertEquals(401, $response->getStatusCode());
+        $this->assertEquals("", $response->getBody());
+        $this->assertEquals("https://example.com/api/foo?bar=pop", $dummy);
+    }
+
+    public function testShouldUseCookieIfHeaderMissingIssue156()
+    {
+        $request = (new ServerRequestFactory)
+            ->createServerRequest("GET", "https://example.com/api")
+            ->withCookieParams(["token" => self::$acmeToken]);
+
+        $default = function (ServerRequestInterface $request) {
+            $response = (new ResponseFactory)->createResponse();
+            $response->getBody()->write("Success");
+            return $response;
+        };
+
+        $collection = new MiddlewareCollection([
+            new JwtAuthentication([
+                "secret" => "supersecretkeyyoushouldnotcommittogithub",
+                "header" => "X-Token",
+                "regexp" => "/(.*)/",
+            ])
+        ]);
+
+        $response = $collection->dispatch($request, $default);
 
         $this->assertEquals(200, $response->getStatusCode());
         $this->assertEquals("Success", $response->getBody());